requirements.txt
langgraph>=0.2.0
langchain-groq>=0.1.0
langchain-core>=0.3.0
tavily-python>=0.3.0
boto3>=1.34.0
pytz>=2024.1
python-dotenv>=1.0.0
pydantic>=2.5.0
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-asyncio>=0.21.0